Moved PostStyles to its own file
