+++ components/posts/Post.js
@@ -1,61 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { Text, View, Image, Dimensions, StyleSheet } from 'react-native';
+import { Text, View, Image } from 'react-native';
 import Interactions from './Interactions';
+import PostStyles from '../PostStyles';
 import axios from 'axios';
-
-const window = Dimensions.get('window');
-const windowWidth = window.width;
-
-// Styles useful for posts. (probably move to independent file soon?)
-const PostStyles = StyleSheet.create ({
-  container: {
-    display: 'flex',
-    justifyContent: 'flex-start',
-  },
-  flexHori: {
-    flexDirection: 'row',
-  },
-  flexVert: {
-    flexDirection: 'column',
-  },
-  topBorder: {
-    borderTopWidth: 1,
-    borderTopColor: '#888888',
-  },
-  padded: {
-    padding: 15,
-  },
-  user: {
-    color: 'rgb(0,0,0)',
-    fontSize: 13,
-    fontWeight: 'bold',
-  },
-  text: {
-    color: 'rgb(0,0,0)',
-    fontSize: 25,
-    fontWeight: '100',
-    textAlign: 'center',
-    marginLeft: 10,
-  },
-  header: {
-    marginTop: 25,
-    padding: 5,
-    paddingLeft: 10,
-    alignItems: 'flex-start',
-  },
-  image: {
-    width: 40,
-    height: 40,
-    borderRadius: 25,
-  },
-  body: {
-    marginTop: 4,
-    width: (windowWidth - 100),
-  },
-  negativeMargin: {
-    marginTop: -15,
-  },
-});
 
 const Post = ({ data }) => {
   const [profilePhoto, setProfilePhoto] = useState(undefined);
